requirements.txt
Django>=3.2,<5
djangorestframework
